Gitlab-chatbot/requirements.txt
google-genai
supabase
sentence-transformers
requests
beautifulsoup4
python-dotenv